Moves Pebble plan types and helpers to public.
This change moves around Pebble types and utils associated with plans
into public packages (not internal). This change was done so
that other software configuring Pebble plans can have accesss to the
concrete types to build up configuration i.e Juju.
