+++ internal/overlord/checkstate/manager.go
@@ -21,7 +21,7 @@ import (
 	"time"
 
 	"github.com/canonical/pebble/internal/logger"
-	"github.com/canonical/pebble/internal/plan"
+	"github.com/canonical/pebble/plan"
 )
 
 // CheckManager starts and manages the health checks.
