+++ internal/overlord/servstate/manager.go
@@ -12,9 +12,9 @@ import (
 	"github.com/canonical/pebble/internal/logger"
 	"github.com/canonical/pebble/internal/overlord/restart"
 	"github.com/canonical/pebble/internal/overlord/state"
-	"github.com/canonical/pebble/internal/plan"
 	"github.com/canonical/pebble/internal/reaper"
 	"github.com/canonical/pebble/internal/servicelog"
+	"github.com/canonical/pebble/plan"
 )
 
 type ServiceManager struct {
