+++ internal/sortutil/tarjan_test.go
@@ -24,10 +24,12 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-package plan
+package sortutil_test
 
 import (
 	. "gopkg.in/check.v1"
+
+	"github.com/canonical/pebble/internal/sortutil"
 )
 
 type TarjanSuite struct{}
@@ -47,7 +49,7 @@ func (TarjanSuite) TestExample(c *C) {
 		"9": {},
 	}
 
-	c.Assert(tarjanSort(successors), DeepEquals, [][]string{
+	c.Assert(sortutil.TarjanSort(successors), DeepEquals, [][]string{
 		{"9"},
 		{"6", "7", "8"},
 		{"5"},
