+++ plan/plan.go
@@ -29,6 +29,7 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/canonical/pebble/internal/osutil"
+	"github.com/canonical/pebble/internal/sortutil"
 	"github.com/canonical/pebble/internal/strutil/shlex"
 )
 
@@ -685,7 +686,7 @@ func order(services map[string]*Service, names []string, stop bool) ([]string, e
 
 	// Sort them up.
 	var order []string
-	for _, names := range tarjanSort(successors) {
+	for _, names := range sortutil.TarjanSort(successors) {
 		if len(names) > 1 {
 			return nil, &FormatError{
 				Message: fmt.Sprintf("services in before/after loop: %s", strings.Join(names, ", ")),
